+++ cmd/notation/verify.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"context"
 	"errors"
 	"fmt"
 	"math"
@@ -9,17 +8,13 @@ import (
 	"reflect"
 
 	"github.com/notaryproject/notation-go"
-	"github.com/notaryproject/notation-go/log"
-	notationregistry "github.com/notaryproject/notation-go/registry"
 	"github.com/notaryproject/notation-go/verifier"
 	"github.com/notaryproject/notation-go/verifier/trustpolicy"
 	"github.com/notaryproject/notation/internal/cmd"
 	"github.com/notaryproject/notation/internal/ioutil"
-	"github.com/opencontainers/go-digest"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 
 	"github.com/spf13/cobra"
-	"oras.land/oras-go/v2/registry"
 )
 
 const maxSignatureAttempts = math.MaxInt64
@@ -30,7 +25,7 @@ type verifyOpts struct {
 	reference        string
 	pluginConfig     []string
 	userMetadata     []string
-	localContent     bool
+	ociLayout        bool
 	trustPolicyScope string
 }
 
@@ -72,9 +67,9 @@ Example - Verify a signature on an OCI artifact identified by a tag and referenc
 	opts.SecureFlagOpts.ApplyFlags(command.Flags())
 	command.Flags().StringArrayVar(&opts.pluginConfig, "plugin-config", nil, "{key}={value} pairs that are passed as it is to a plugin, if the verification is associated with a verification plugin, refer plugin documentation to set appropriate values")
 	cmd.SetPflagUserMetadata(command.Flags(), &opts.userMetadata, cmd.PflagUserMetadataVerifyUsage)
-	command.Flags().BoolVar(&opts.localContent, "local-content", false, "verify local artifact")
-	command.Flags().StringVar(&opts.trustPolicyScope, "scope", "", "trust policy scope for local artifact verification. This flag is required when local-content is set to true")
-	command.MarkFlagsRequiredTogether("local-content", "scope")
+	command.Flags().BoolVar(&opts.ociLayout, "oci-layout", false, "verify local content")
+	command.Flags().StringVar(&opts.trustPolicyScope, "scope", "", "trust policy scope for local content verification")
+	command.MarkFlagsRequiredTogether("oci-layout", "scope")
 	return command
 }
 
@@ -101,94 +96,123 @@ func runVerify(command *cobra.Command, opts *verifyOpts) error {
 	}
 
 	// core verify process
-	if opts.localContent {
-		printOut, outcomes, err := verifyLocal(ctx, opts, verifier, configs, userMetadata)
-		if err != nil {
-			return err
-		}
-		onSucess(outcomes, printOut)
-	} else {
-		printOut, outcomes, err := verifyRemote(ctx, opts, verifier, configs, userMetadata)
-		if err != nil {
-			return err
-		}
-		onSucess(outcomes, printOut)
+	var inputType inputType = remoteRegistry
+	if opts.ociLayout {
+		inputType = ociLayout
+		// printOut, outcomes, err := verifyLocal(ctx, opts, verifier, configs, userMetadata)
+		// if err != nil {
+		// 	return err
+		// }
+		// onSucess(outcomes, printOut)
+		// } else {
+		// 	printOut, outcomes, err := verifyRemote(ctx, opts, verifier, configs, userMetadata)
+		// 	if err != nil {
+		// 		return err
+		// 	}
+		// 	onSucess(outcomes, printOut)
+		// }
 	}
-	return nil
-}
-
-func verifyRemote(ctx context.Context, opts *verifyOpts, verifier notation.Verifier, configs, userMetadata map[string]string) (string, []*notation.VerificationOutcome, error) {
 	reference := opts.reference
-	sigRepo, err := getRemoteRepository(ctx, &opts.SecureFlagOpts, reference)
+	sigRepo, err := getRepository(ctx, inputType, reference, &opts.SecureFlagOpts)
 	if err != nil {
-		return "", nil, err
+		return err
 	}
 	// resolve the given reference and set the digest
-	ref, err := resolveReference(ctx, &opts.SecureFlagOpts, reference, sigRepo, func(ref registry.Reference, manifestDesc ocispec.Descriptor) {
-		fmt.Fprintf(os.Stderr, "Warning: Always verify the artifact using digest(@sha256:...) rather than a tag(:%s) because resolved digest may not point to the same signed artifact, as tags are mutable.\n", ref.Reference)
+	_, fullRef, err := resolveReference(ctx, inputType, reference, sigRepo, func(ref string, manifestDesc ocispec.Descriptor) {
+		fmt.Fprintf(os.Stderr, "Warning: Always verify the artifact using digest(@sha256:...) rather than a tag(:%s) because resolved digest may not point to the same signed artifact, as tags are mutable.\n", ref)
 	})
 	if err != nil {
-		return "", nil, err
+		return err
 	}
 	verifyOpts := notation.VerifyOptions{
-		ArtifactReference: ref.String(),
+		ArtifactReference: fullRef,
 		PluginConfig:      configs,
 		// TODO: need to change MaxSignatureAttempts as a user input flag or
 		// a field in config.json
 		MaxSignatureAttempts: maxSignatureAttempts,
 		UserMetadata:         userMetadata,
 	}
-
-	// core verify process
 	_, outcomes, err := notation.Verify(ctx, verifier, sigRepo, verifyOpts)
-	err = checkFailure(outcomes, ref.String(), err)
+	err = checkFailure(outcomes, fullRef, err)
 	if err != nil {
-		return "", nil, err
+		return err
 	}
-	return ref.String(), outcomes, nil
+	onSucess(outcomes, fullRef)
+	return nil
 }
 
-func verifyLocal(ctx context.Context, opts *verifyOpts, verifier notation.Verifier, configs, userMetadata map[string]string) (string, []*notation.VerificationOutcome, error) {
-	layoutPath, layoutReference, err := parseOCILayoutReference(opts.reference)
-	if err != nil {
-		return "", nil, err
-	}
-	return verifyFromFolder(ctx, opts, verifier, layoutPath, layoutReference, configs, userMetadata)
-}
+// func verifyRemote(ctx context.Context, opts *verifyOpts, verifier notation.Verifier, configs, userMetadata map[string]string) (string, []*notation.VerificationOutcome, error) {
+// 	reference := opts.reference
+// 	sigRepo, err := getRemoteRepository(ctx, &opts.SecureFlagOpts, reference)
+// 	if err != nil {
+// 		return "", nil, err
+// 	}
+// 	// resolve the given reference and set the digest
+// 	ref, err := resolveReference(ctx, &opts.SecureFlagOpts, reference, sigRepo, func(ref registry.Reference, manifestDesc ocispec.Descriptor) {
+// 		fmt.Fprintf(os.Stderr, "Warning: Always verify the artifact using digest(@sha256:...) rather than a tag(:%s) because resolved digest may not point to the same signed artifact, as tags are mutable.\n", ref.Reference)
+// 	})
+// 	if err != nil {
+// 		return "", nil, err
+// 	}
+// 	verifyOpts := notation.VerifyOptions{
+// 		ArtifactReference: ref.String(),
+// 		PluginConfig:      configs,
+// 		// TODO: need to change MaxSignatureAttempts as a user input flag or
+// 		// a field in config.json
+// 		MaxSignatureAttempts: maxSignatureAttempts,
+// 		UserMetadata:         userMetadata,
+// 	}
 
-func verifyFromFolder(ctx context.Context, opts *verifyOpts, verifier notation.Verifier, path, reference string, configs, userMetadata map[string]string) (string, []*notation.VerificationOutcome, error) {
-	logger := log.GetLogger(ctx)
+// 	// core verify process
+// 	_, outcomes, err := notation.Verify(ctx, verifier, sigRepo, verifyOpts)
+// 	err = checkFailure(outcomes, ref.String(), err)
+// 	if err != nil {
+// 		return "", nil, err
+// 	}
+// 	return ref.String(), outcomes, nil
+// }
 
-	sigRepo, err := notationregistry.NewOCIRepository(path, notationregistry.RepositoryOptions{})
-	if err != nil {
-		return "", nil, err
-	}
-	targetDesc, err := sigRepo.Resolve(ctx, reference)
-	if err != nil {
-		return "", nil, fmt.Errorf("failed to resolve OCI layout reference: %s", err)
-	}
-	logger.Infof("OCI layout reference %s resolved to target manifest descriptor: %+v", reference, targetDesc)
-	if digest.Digest(reference).Validate() != nil {
-		// layout.reference is a tag
-		fmt.Fprintf(os.Stderr, "Warning: Always verify the artifact using digest(@sha256:...) rather than a tag(:%s) because resolved digest may not point to the same signed artifact, as tags are mutable.\n", reference)
-	}
-	reference = targetDesc.Digest.String()
-	printOut := path + "@" + targetDesc.Digest.String()
-	verifyOpts := notation.VerifyOptions{
-		ArtifactReference:    opts.trustPolicyScope + "@" + reference,
-		PluginConfig:         configs,
-		MaxSignatureAttempts: maxSignatureAttempts,
-		UserMetadata:         userMetadata,
-	}
+// func verifyLocal(ctx context.Context, opts *verifyOpts, verifier notation.Verifier, configs, userMetadata map[string]string) (string, []*notation.VerificationOutcome, error) {
+// 	layoutPath, layoutReference, err := parseOCILayoutReference(opts.reference)
+// 	if err != nil {
+// 		return "", nil, err
+// 	}
+// 	return verifyFromFolder(ctx, opts, verifier, layoutPath, layoutReference, configs, userMetadata)
+// }
 
-	// core process
-	_, outcomes, err := notation.Verify(ctx, verifier, sigRepo, verifyOpts)
-	err = checkFailure(outcomes, printOut, err)
-	if err != nil {
-		return "", nil, err
-	}
-	return printOut, outcomes, nil
-}
+// func verifyFromFolder(ctx context.Context, opts *verifyOpts, verifier notation.Verifier, path, reference string, configs, userMetadata map[string]string) (string, []*notation.VerificationOutcome, error) {
+// 	logger := log.GetLogger(ctx)
+
+// 	sigRepo, err := notationregistry.NewOCIRepository(path, notationregistry.RepositoryOptions{})
+// 	if err != nil {
+// 		return "", nil, err
+// 	}
+// 	targetDesc, err := sigRepo.Resolve(ctx, reference)
+// 	if err != nil {
+// 		return "", nil, fmt.Errorf("failed to resolve OCI layout reference: %s", err)
+// 	}
+// 	logger.Infof("OCI layout reference %s resolved to target manifest descriptor: %+v", reference, targetDesc)
+// 	if digest.Digest(reference).Validate() != nil {
+// 		// layout.reference is a tag
+// 		fmt.Fprintf(os.Stderr, "Warning: Always verify the artifact using digest(@sha256:...) rather than a tag(:%s) because resolved digest may not point to the same signed artifact, as tags are mutable.\n", reference)
+// 	}
+// 	reference = targetDesc.Digest.String()
+// 	printOut := path + "@" + targetDesc.Digest.String()
+// 	verifyOpts := notation.VerifyOptions{
+// 		ArtifactReference:    opts.trustPolicyScope + "@" + reference,
+// 		PluginConfig:         configs,
+// 		MaxSignatureAttempts: maxSignatureAttempts,
+// 		UserMetadata:         userMetadata,
+// 	}
+
+// 	// core process
+// 	_, outcomes, err := notation.Verify(ctx, verifier, sigRepo, verifyOpts)
+// 	err = checkFailure(outcomes, printOut, err)
+// 	if err != nil {
+// 		return "", nil, err
+// 	}
+// 	return printOut, outcomes, nil
+// }
 
 func checkFailure(outcomes []*notation.VerificationOutcome, printOut string, err error) error {
 	// write out on failure
